+++ tests/unit_tests/model_tests/test_property_statistics_model.py
@@ -7,10 +7,12 @@ from webviz_subsurface.plugins._property_statistics.models.property_statistics_m
     PropertyStatisticsModel,
 )
 
+from webviz_subsurface._providers import EnsembleTableProviderFactory
 
-def get_data_df(testdata_folder: Path) -> pd.DataFrame:
 
-    return pd.read_csv(
+def get_provider(testdata_folder: Path) -> pd.DataFrame:
+    table_provider = EnsembleTableProviderFactory.instance()
+    return table_provider.create_provider_set_from_aggregated_csv_file(
         testdata_folder
         / "reek_test_data"
         / "aggregated_data"
@@ -19,8 +21,9 @@ def get_data_df(testdata_folder: Path) -> pd.DataFrame:
 
 
 def test_init(testdata_folder: Path) -> None:
-    data_df = get_data_df(testdata_folder)
-    model = PropertyStatisticsModel(dataframe=data_df, theme=default_theme)
+
+    provider = get_provider(testdata_folder)
+    model = PropertyStatisticsModel(provider=provider, theme=default_theme)
     assert set(model.dataframe.columns) == set(
         [
             "PROPERTY",