+++ webviz_subsurface/plugins/_parameter_analysis/parameter_analysis.py
@@ -1,121 +1,137 @@
-import pathlib
-from typing import Optional
+from pathlib import Path
+from typing import Optional, Tuple
 
 from dash import dcc
 from webviz_config import WebvizPluginABC, WebvizSettings
+from webviz_config.deprecation_decorators import deprecated_plugin_arguments
 
-from webviz_subsurface._models import (
-    EnsembleSetModel,
-    caching_ensemble_set_model_factory,
+from webviz_subsurface._providers import (
+    EnsembleSummaryProviderFactory,
+    EnsembleTableProviderFactory,
+    Frequency,
 )
 
 from .controllers import parameter_qc_controller, parameter_response_controller
 from .data_loaders import read_csv
-from .models import ParametersModel, SimulationTimeSeriesModel
+from .models import (
+    ParametersModel,
+    ProviderTimeSeriesDataModel,
+    SimulationTimeSeriesModel,
+)
 from .views import main_view
 
 
+def check_deprecation_argument(
+    csvfile_parameters: Optional[Path], csvfile_smry: Optional[Path]
+) -> Optional[Tuple[str, str]]:
+    if any(elm is not None for elm in [csvfile_parameters, csvfile_smry]):
+        return (
+            "The usage of aggregated csvfiles as user input options are deprecated. "
+            "Please provide feedback if you see a need for a continuation "
+            "of this functionality ",
+            "",
+        )
+    return None
+
+
 class ParameterAnalysis(WebvizPluginABC):
-    """This plugin visualizes parameter distributions and statistics. /
-    for FMU ensembles, and can be used to investigate parameter correlations /
-    on reservoir simulation time series data.
+    """This plugin visualizes parameter distributions and statistics.
+for FMU ensembles, and can be used to investigate parameter correlations
+on reservoir simulation time series data.
 
 ---
 
-**Input data can be provided in two ways: Aggregated or read from ensembles stored on scratch.**
-
-**Using aggregated data**
-* **`csvfile_parameters`:** Aggregated `csv` file with `REAL`, `ENSEMBLE` and parameter columns. \
-    (absolute path or relative to config file).
-* **`csvfile_smry`:** (Optional) Aggregated `csv` file for volumes with `REAL`, `ENSEMBLE`, `DATE` \
-    and vector columns (absolute path or relative to config file).
-
 **Using raw ensemble data stored in realization folders**
 * **`ensembles`:** Which ensembles in `shared_settings` to visualize.
-* **`column_keys`:** List of vectors to extract. If not given, all vectors \
-    from the simulations will be extracted. Wild card asterisk `*` can be used.
+* **`rel_file_pattern`:** path to `.arrow` files with summary data.
 * **`time_index`:** Time separation between extracted values. Can be e.g. `monthly` (default) or \
     `yearly`.
-
-**Common settings for all input options**
 * **`drop_constants`:** Bool used to determine if constant parameters should be dropped. \
     Default is True.
 
 ---
 
+?> `Arrow` format for simulation time series data can be generated using the `ECL2CSV` forward \
+model in ERT. On existing ensembles the command line tool `smry2arrow_batch` can be used to \
+generate arrow files.
+
 !> For smry data it is **strongly recommended** to keep the data frequency to a regular frequency \
-(like `monthly` or `yearly`). This applies to both csv input and when reading from `UNSMRY` \
-(controlled by the `sampling` key). This is because the statistics and fancharts are calculated \
+(like `monthly` or `yearly`). This is because the statistics are calculated \
 per DATE over all realizations in an ensemble, and the available dates should therefore not \
 differ between individual realizations of an ensemble.
 
 ?> Vectors that are identified as historical vectors (e.g. FOPTH is the history of FOPT) will \
 be plotted together with their non-historical counterparts as reference lines.
 
-**Using simulation time series data directly from `.UNSMRY` files**
-
 !> Parameter values are extracted automatically from the `parameters.txt` files in the individual
-realizations if you have defined `ensembles`, using the `fmu-ensemble` library.
+realizations if you have defined `ensembles`.
 
-!> The `UNSMRY` files are auto-detected by `fmu-ensemble` in the `eclipse/model` folder of the \
-individual realizations. You should therefore not have more than one `UNSMRY` file in this \
-folder, to avoid risk of not extracting the right data.
-
-**Using aggregated data**
-
-?> Aggregated data may speed up the build of the app, as processing of `UNSMRY` files can be \
-slow for large models.
 
 """
 
-    # pylint: disable=too-many-arguments
+    # pylint: disable=too-many-arguments, too-many-locals
+    @deprecated_plugin_arguments(check_deprecation_argument)
     def __init__(
         self,
         app,
         webviz_settings: WebvizSettings,
         ensembles: Optional[list] = None,
-        csvfile_parameters: pathlib.Path = None,
-        csvfile_smry: pathlib.Path = None,
         time_index: str = "monthly",
-        column_keys: Optional[list] = None,
         drop_constants: bool = True,
+        rel_file_pattern: str = "share/results/unsmry/*.arrow",
+        csvfile_parameters: Optional[Path] = None,
+        csvfile_smry: Optional[Path] = None,
     ):
         super().__init__()
 
         self.theme = webviz_settings.theme
-        self.time_index = time_index
-        self.column_keys = column_keys
         self.ensembles = ensembles
         self.csvfile_parameters = csvfile_parameters
         self.csvfile_smry = csvfile_smry
 
+        table_provider = EnsembleTableProviderFactory.instance()
+
         if ensembles is not None:
-            self.emodel: EnsembleSetModel = (
-                caching_ensemble_set_model_factory.get_or_create_model(
-                    ensemble_paths={
-                        ens: webviz_settings.shared_settings["scratch_ensembles"][ens]
-                        for ens in ensembles
+            ensemble_paths = {
+                ensemble_name: webviz_settings.shared_settings["scratch_ensembles"][
+                    ensemble_name
+                ]
+                for ensemble_name in ensembles
+            }
+
+            resampling_frequency = Frequency(time_index)
+            provider_factory = EnsembleSummaryProviderFactory.instance()
+
+            try:
+                self.vmodel = ProviderTimeSeriesDataModel(
+                    provider_set={
+                        ens: provider_factory.create_from_arrow_unsmry_presampled(
+                            str(ens_path), rel_file_pattern, resampling_frequency
+                        )
+                        for ens, ens_path in ensemble_paths.items()
                     },
-                    time_index=self.time_index,
-                    column_keys=self.column_keys,
                 )
-            )
-            self.pmodel = ParametersModel(
-                dataframe=self.emodel.load_parameters(),
-                theme=self.theme,
-                drop_constants=drop_constants,
-            )
-            self.vmodel = SimulationTimeSeriesModel(
-                dataframe=self.emodel.get_or_load_smry_cached()
+            except ValueError as error:
+                message = (
+                    f"No arrow files found at {rel_file_pattern}. If no arrow files have been "
+                    "generated with `ERT` using `ECL2CSV`, the commandline tool "
+                    "`smry2arrow_batch` can be used to generate arrow files for an ensemble"
+                )
+                raise ValueError(message) from error
+
+            parameterproviderset = (
+                table_provider.create_provider_set_from_per_realization_parameter_file(
+                    ensemble_paths
+                )
             )
 
         elif self.csvfile_parameters is None:
             raise ValueError("Either ensembles or csvfile_parameters must be specified")
         else:
-            self.pmodel = ParametersModel(
-                dataframe=read_csv(csvfile_parameters),
-                theme=self.theme,
-                drop_constants=drop_constants,
+            parameterproviderset = (
+                table_provider.create_provider_set_from_aggregated_csv_file(
+                    csvfile_parameters
+                )
             )
             if self.csvfile_smry is not None:
                 self.vmodel = SimulationTimeSeriesModel(
@@ -123,6 +139,12 @@ slow for large models.
                 )
             else:
                 self.vmodel = None
+
+        self.pmodel = ParametersModel(
+            provider=parameterproviderset,
+            theme=self.theme,
+            drop_constants=drop_constants,
+        )
 
         self.set_callbacks(app)
 
@@ -147,9 +169,7 @@ slow for large models.
 
     def add_webvizstore(self):
         store = []
-        if self.ensembles is not None:
-            store.extend(self.emodel.webvizstore)
-        else:
+        if self.ensembles is None:
             store.append(
                 (
                     read_csv,