+++ webviz_subsurface/plugins/_property_statistics/controllers/property_response_controller.py
@@ -1,16 +1,25 @@
 from typing import Callable, Tuple, Union
 
-import numpy as np
 import pandas as pd
 from dash import ALL, Dash, Input, Output, State, callback_context, no_update
 from dash.dash import _NoUpdate
 from dash.exceptions import PreventUpdate
 
+from webviz_subsurface._figures import BarChart, ScatterPlot, TimeSeriesFigure
 from webviz_subsurface._models import SurfaceLeafletModel
+from webviz_subsurface._utils.dataframe_utils import (
+    correlate_response_with_dataframe,
+    merge_dataframes_on_realization,
+)
+from webviz_subsurface.plugins._simulation_time_series.utils.datetime_utils import (
+    from_str,
+)
 
-from ..figures.correlation_figure import CorrelationFigure
-from ..models import PropertyStatisticsModel, SimulationTimeSeriesModel
-from ..utils.colors import find_intermediate_color_rgba
+from ..models import (
+    PropertyStatisticsModel,
+    ProviderTimeSeriesDataModel,
+    SimulationTimeSeriesModel,
+)
 from ..utils.surface import surface_from_zone_prop
 
 
@@ -18,7 +27,7 @@ def property_response_controller(
     get_uuid: Callable,
     surface_table: pd.DataFrame,
     property_model: PropertyStatisticsModel,
-    timeseries_model: SimulationTimeSeriesModel,
+    timeseries_model: Union[SimulationTimeSeriesModel, ProviderTimeSeriesDataModel],
     app: Dash,
 ) -> None:
     @app.callback(
@@ -99,8 +108,9 @@ def property_response_controller(
     @app.callback(
         Output(get_uuid("property-response-vector-graph"), "figure"),
         Output(get_uuid("property-response-correlation-graph"), "figure"),
+        Output(get_uuid("property-response-scatter-graph"), "figure"),
         Input({"id": get_uuid("ensemble-selector"), "tab": "response"}, "value"),
-        Input(get_uuid("property-response-vector-select"), "value"),
+        Input(get_uuid("property-response-vector-select"), "selectedNodes"),
         Input(get_uuid("property-response-vector-graph"), "clickData"),
         Input(get_uuid("property-response-correlation-graph"), "clickData"),
         Input(
@@ -125,14 +135,17 @@ def property_response_controller(
         corr_filter: list,
         figure: dict,
         label: str,
-    ) -> Tuple[dict, dict]:
+    ) -> Tuple[dict, dict, dict]:
+
         if (
             callback_context.triggered is None
             or callback_context.triggered[0]["prop_id"] == "."
             or vector is None
+            or not all(filt for filt in selectors)
         ):
             raise PreventUpdate
-        ctx = callback_context.triggered[0]["prop_id"].split(".")[0]
+
+        vector = vector[0]
 
         # Filter realizations if correlation filter is active
         real_filter = (
@@ -140,20 +153,6 @@ def property_response_controller(
             if corr_filter[0] is not None and corr_filter[1] is not None
             else None
         )
-
-        # Make timeseries graph
-        if any(
-            substr in ctx
-            for substr in [
-                get_uuid("property-response-vector-select"),
-                get_uuid("ensemble-selector"),
-                get_uuid("property-response-correlated-slider"),
-            ]
-        ):
-
-            figure = update_timeseries_graph(
-                timeseries_model, ensemble, vector, real_filter
-            )
 
         # Get clicked data or last available date initially
         date = (
@@ -163,33 +162,38 @@ def property_response_controller(
             if timeseries_clickdata
             else timeseries_model.get_last_date(ensemble)
         )
-
-        # Draw clicked date as a black line
-        ymin = min([min(trace["y"]) for trace in figure["data"]])
-        ymax = max([max(trace["y"]) for trace in figure["data"]])
-        figure["layout"]["shapes"] = [
-            {"type": "line", "x0": date, "x1": date, "y0": ymin, "y1": ymax}
-        ]
+        date = from_str(date) if isinstance(date, str) else date
 
         # Get dataframe with vector and REAL
-        vector_df = timeseries_model.get_ensemble_vector_for_date(
-            ensemble=ensemble, vector=vector, date=date
+        vector_df = timeseries_model.get_vector_df(
+            ensemble=ensemble, vectors=[vector], realizations=real_filter
         )
-        vector_df["REAL"] = vector_df["REAL"].astype(int)
 
         # Get dataframe with properties per label and REAL
         prop_df = property_model.get_ensemble_properties(ensemble, selectors)
-        prop_df["REAL"] = prop_df["REAL"].astype(int)
         prop_df = (
             prop_df[prop_df["REAL"].isin(real_filter)]
             if real_filter is not None
             else prop_df
         )
 
+        merged_df = merge_dataframes_on_realization(
+            dframe1=vector_df[vector_df["DATE"] == date], dframe2=prop_df
+        )
         # Correlate properties against vector
-        corrseries = correlate(vector_df, prop_df, response=vector)
+        corrseries = correlate_response_with_dataframe(
+            merged_df,
+            response=vector,
+            corrwith=[col for col in prop_df if col != "REAL"],
+        )
+        # Handle missing
+        if corrseries.empty:
+            return {}, {}, {}
+
         # Make correlation figure
-        correlation_figure = CorrelationFigure(corrseries, n_rows=20, title="")
+        correlation_figure = BarChart(
+            corrseries, n_rows=15, title=f"Correlations with {vector}", orientation="h"
+        )
 
         # Get clicked correlation bar or largest bar initially
         selected_corr = (
@@ -199,72 +203,30 @@ def property_response_controller(
         )
 
         # Update bar colors
-        correlation_figure.set_bar_colors(selected_corr)
-
-        # Order realizations sorted on value of property
-        real_order = (
-            property_model.get_real_order(ensemble, series=selected_corr)
-            if selected_corr is not None
-            else None
+        correlation_figure.color_bars(selected_corr, color="#007079", opacity=0.5)
+        prop_df_norm = property_model.get_real_and_value_df(
+            ensemble, series=selected_corr, normalize=True
         )
+        figure = TimeSeriesFigure(
+            dframe=merge_dataframes_on_realization(vector_df, prop_df_norm),
+            visualization="realizations",
+            vector=vector,
+            ensemble=ensemble,
+            dateline=date,
+            historical_vector_df=timeseries_model.get_historical_vector_df(
+                vector, ensemble
+            ),
+            color_col=selected_corr,
+            line_shape_fallback=timeseries_model.line_shape_fallback,
+        ).figure
 
-        # Color timeseries lines from value of property
-        if real_order is not None:
-            mean = real_order["Avg"].mean()
-            low_reals = real_order[real_order["Avg"] <= mean]["REAL"].astype(str).values
-            high_reals = real_order[real_order["Avg"] > mean]["REAL"].astype(str).values
-            for trace_no, trace in enumerate(figure.get("data", [])):
-                if trace["name"] == ensemble:
-                    figure["data"][trace_no]["marker"]["color"] = set_real_color(
-                        str(trace["customdata"]), low_reals, high_reals
-                    )
-            figure["layout"]["title"] = f"Colored by {selected_corr}"
+        scatter_fig = ScatterPlot(
+            merged_df if selected_corr in merged_df else pd.DataFrame(),
+            response=vector,
+            param=selected_corr,
+            color="#007079",
+            title=f"{vector} vs {selected_corr}",
+            plot_trendline=True,
+        ).figure
 
-        return figure, correlation_figure.figure
-
-
-def set_real_color(real_no: str, low_reals: list, high_reals: list) -> str:
-
-    if real_no in low_reals:
-        index = int(list(low_reals).index(real_no))
-        intermed = index / len(low_reals)
-        return find_intermediate_color_rgba(
-            "rgba(255,0,0, 100, .1)", "rgba(220,220,220, 0.1)", intermed
-        )
-    if real_no in high_reals:
-        index = int(list(high_reals).index(real_no))
-        intermed = index / len(high_reals)
-        return find_intermediate_color_rgba(
-            "rgba(220,220,220, 0.1)", "rgba(50,205,50, 1)", intermed
-        )
-
-    return "rgba(220,220,220, 0.2)"
-
-
-def update_timeseries_graph(
-    timeseries_model: SimulationTimeSeriesModel,
-    ensemble: str,
-    vector: str,
-    real_filter: pd.Series = None,
-) -> dict:
-
-    return {
-        "data": timeseries_model.add_realization_traces(
-            ensemble=ensemble, vector=vector, real_filter=real_filter
-        ),
-        "layout": dict(
-            margin={"r": 40, "l": 40, "t": 40, "b": 40},
-        ),
-    }
-
-
-def correlate(vectordf: pd.DataFrame, propdf: pd.DataFrame, response: str) -> pd.Series:
-    """Returns the correlation matrix for a dataframe"""
-    df = pd.merge(propdf, vectordf, on=["REAL"])
-    df = df[df.columns[df.nunique() > 1]]
-    if response not in df.columns:
-        df[response] = np.nan
-    series = df[response]
-    df = df.drop(columns=[response, "REAL"])
-    corrdf = df.corrwith(series)
-    return corrdf.reindex(corrdf.abs().sort_values().index)
+        return figure, correlation_figure.figure, scatter_fig