+++ webviz_subsurface/plugins/_property_statistics/views/main_view.py
@@ -1,9 +1,12 @@
-from typing import Callable, Dict, List, Optional
+from typing import Callable, Dict, Optional, Union
 
 import webviz_core_components as wcc
-from dash import dcc, html
 
-from ..models import PropertyStatisticsModel
+from ..models import (
+    PropertyStatisticsModel,
+    ProviderTimeSeriesDataModel,
+    SimulationTimeSeriesModel,
+)
 from .property_delta_view import property_delta_view
 from .property_qc_view import property_qc_view
 from .property_response_view import property_response_view
@@ -12,18 +15,20 @@ from .property_response_view import property_response_view
 def main_view(
     get_uuid: Callable,
     property_model: PropertyStatisticsModel,
-    vector_options: List[Dict],
+    vector_model: Optional[
+        Union[SimulationTimeSeriesModel, ProviderTimeSeriesDataModel]
+    ],
     surface_folders: Optional[Dict] = None,
-) -> dcc.Tabs:
+) -> wcc.Tabs:
     tabs = [
-        make_tab(
+        wcc.Tab(
             label="Property QC",
             children=property_qc_view(get_uuid=get_uuid, property_model=property_model),
         )
     ]
     if len(property_model.ensembles) > 1:
         tabs.append(
-            make_tab(
+            wcc.Tab(
                 label="AHM impact on property",
                 children=property_delta_view(
                     get_uuid=get_uuid,
@@ -32,30 +37,17 @@ def main_view(
                 ),
             )
         )
-    tabs.append(
-        make_tab(
-            label="Property impact on simulation profiles",
-            children=property_response_view(
-                get_uuid=get_uuid,
-                property_model=property_model,
-                vector_options=vector_options,
-                surface_folders=surface_folders,
+    if vector_model is not None:
+        tabs.append(
+            wcc.Tab(
+                label="Property impact on simulation profiles",
+                children=property_response_view(
+                    get_uuid=get_uuid,
+                    property_model=property_model,
+                    vector_model=vector_model,
+                    surface_folders=surface_folders,
+                ),
             ),
-        ),
-    )
+        )
 
-    return html.Div(
-        id=get_uuid("layout"),
-        children=wcc.Tabs(
-            style={"width": "100%"},
-            children=tabs,
-        ),
-    )
-
-
-def make_tab(label: str, children: wcc.FlexBox) -> dcc.Tab:
-
-    return wcc.Tab(
-        label=label,
-        children=children,
-    )
+    return wcc.Tabs(style={"width": "100%"}, children=tabs)